some changes
move player licenses to config so admins don't have to configure this item in the code.

server/functions made it so that playerdata pays through the payment variable before it searches the default payrate in the jobs config.

player.lua
made jobrep a thing for all jobs
added jobs, gangs, gangrep, and jobhistory to the metadata.

moved license metadata defaults to be configured in config.lua

reworked plan with playerdata.jobs and moved it to metadata

as was done in the qb-jobs config made the grades a number instead of a string.

added quit, and fired flags to the job name to ensure that quitting and termination is being recored.

added recording of promotions and demotions

made it so that the payrate is recored in the playerdata to allow for raises and cuts

added jobhistory meta data to setjob

as was done with jobs was done with gangs and set the grade to a number instead of a string.

removed the playerdata.jobs that i added earlier.